@@ -1,5 +1,5 @@
 import 'showing.dart';
-import 'Film.dart';
+import 'movie.dart';
 
 class AppState {
   List<Map<String, dynamic>> cart = [];
@@ -7,9 +7,27 @@ class AppState {
   List<Map<String, dynamic>>? reservations;
   DateTime selectedDate =
       DateTime(DateTime.now().year, DateTime.now().month, DateTime.now().day);
-  Film? selectedFilm;
+  Movie? selectedFilm;
   Showing? selectedShowing;
   Map<String, dynamic>? theater;
+
+  AppState copyWith(
+      {List<Map<String, dynamic>>? cart,
+      Customer? customer,
+      List<Map<String, dynamic>>? reservations,
+      DateTime? selectedDate,
+      Movie? selectedFilm,
+      Showing? selectedShowing,
+      Map<String, dynamic>? theater}) {
+    return AppState()
+      ..cart = cart ?? this.cart
+      ..customer = customer ?? this.customer
+      ..reservations = reservations ?? this.reservations
+      ..selectedDate = selectedDate ?? this.selectedDate
+      ..selectedFilm = selectedFilm ?? this.selectedFilm
+      ..selectedShowing = selectedShowing ?? this.selectedShowing
+      ..theater = theater ?? this.theater;
+  }
 }
 
 DateTime now = DateTime.now();
